+++ commands/release/release.go
@@ -43,8 +43,9 @@ const (
 	extraNanosecondsToAddToLastFetchedTimestamp = 0
 	lastFetchedFileMode                         = 0644
 
-	relChangelogFilepath = "docs/changelog.md"
-	relAPIFolderPath     = "api/"
+	relChangelogFilepath                      = "docs/changelog.md"
+	relAPIFolderPathForKurtosisMonorepo       = "api/"
+	typeScriptPackageJSONForMinimalGRPCServer = "typescript/package.json"
 
 	expectedNumTBDHeaderLines         = 1
 	versionToBeReleasedPlaceholderStr = "TBD"
@@ -258,13 +259,14 @@ func run(cmd *cobra.Command, args []string) error {
 	// Commit pre-release changes
 	// These are the only files that should get changed during the pre-release
 	// &git.AddOptions{All: true} ignores git ignore and adds `kurtosis_version`
-	// TODO figure out a way to use All while respecting the ignore for `kurtosis_version`
-	pathsToAdd := []string{relChangelogFilepath, relAPIFolderPath}
+	// TODO figure out a way to use All while respecting the .gitignore for `kurtosis_version`
+	pathsToAdd := []string{relChangelogFilepath, relAPIFolderPathForKurtosisMonorepo, typeScriptPackageJSONForMinimalGRPCServer}
 
 	for _, pathToAdd := range pathsToAdd {
 		err = worktree.AddWithOptions(&git.AddOptions{Glob: pathToAdd})
 		if err != nil {
-			return stacktrace.Propagate(err, "An error occurred while attempting to add '%v' to staging area.", pathsToAdd)
+			logrus.Warnf("An error occurred while adding '%v' does it even exist?", pathToAdd)
+			logrus.Warn("You are probably hitting https://github.com/kurtosis-tech/kudet/issues/22")
 		}
 	}