+++ docs/changelog.md
@@ -1,5 +1,9 @@
 # TBD
 
+### Fixes
+* Adds the package.json changed by the minimal GRPC server during pre-release to staging area
+* Fixes a bug where we try to add a file which doesn't exist that breaks the releae process
+
 # 0.1.8
 
 ### Fixes